requirements.txt
Flask==3.1.0
requests==2.32.3
Pillow==10.3.0
matplotlib==3.9.2
numpy==1.26.4
gunicorn==23.0.0
